split the code
